[aarch64] Allocate LJ_TISNUM early
LJ_TISNUM is allocated too late in the cycle and it ends up reusing a
different register, resulting in a crash under register pressure.
Hoist the allocation to the top so that it is done early enough and
the allowed register set no longer contains that register.  This also
has the nice side effect of beig slightly faster since it hoists a
constant allocation out of the generated loop.