[ppc]  Load BASEP4 as much as possible
BASEP4 doesn't seem to get initialized all the time, especially when
BASE is updated because of which programs can crash at random on
ppc32.  Err on the conservative side and set BASEP4 every time BASE_LO
(or BASE_HI for LE) are accessed.

This eventually needs to be tuned optimally.

Signed-off-by: Siddhesh Poyarekar <siddhesh@gotplt.org>