Fix vulnerability in debug.getinfo
If the first argument is a valid stack level and the first character
in the options argument for debug.getinfo is >, then the final
argument will be used as the function.  e.g.  debug.getinfo(0, ">",
debug.getinfo)

In release mode, the final argument's type is never ensured to be a
function. This means that we can pass whatever we want to it. For
example, if we passed the double representation of a uintptr_t that
points to whichever address we want to read memory from it'll return a
function at that address. To read memory, we can simply call
debug.getfenv on the function and then call string.format("%p",env) to
get the contents of the memory.

This can be further exploited to enable arbitrary memory writing.

Pulled from LuaJIT/LuaJIT#526