fold: Fix up conversion mode in simplify_conv_narrow
Fixes #37.

simplify_conv_narrow transforms a 64-bit to 32-bit int conversion
after an operation (e.g. add, sub, mul) to an op of the converted
operands by simply converting the operands into 32-bit int.  It
however fails to update the destination type of the conversion mode
because of which one ends up with inconsistent IR and fails with an
assertion.

The test case is incorporated into this patch but it doesn't actually
get executed in the make check.  ffi_convert.lua needs extensive
fixing up to enable it in the testsuite and is currently out of scope
for 2.1.2.

Signed-off-by: Siddhesh Poyarekar <siddhesh@gotplt.org>
Signed-off-by: spacewander <spacewanderlzx@gmail.com>