Document release process
moonjit v2.1.1 release was a minor disaster with multiple issues.  One
of the problems was that it was unclear which all files had to be
updated to make the release and therefore a few files in `src/jit`
slipped through without an update.

Document the release process and also all the files that need to be
updated so that there is a ready checklist to follow while making a
release.  Eventually this should just become a script or a make
target.

Signed-off-by: Siddhesh Poyarekar <siddhesh@gotplt.org>